+++ aldrin/src/low_level.rs
@@ -3,9 +3,9 @@
 //! The types in this module are primarily intended for use by the code generator.
 
 mod event;
-mod events;
+mod event_listener;
 
-pub(crate) use events::{EventsId, EventsRequest};
+pub(crate) use event_listener::{EventListenerId, EventListenerRequest};
 
 pub use event::Event;
-pub use events::Events;
+pub use event_listener::EventListener;