+++ aldrin/CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Move `Event` to the `low_level` module.
-- Move `Event` to the `low_level` module.
+- Rename `Events` to `EventListener` and move it to the `low_level` module.
 
 ## [0.4.0] - 2024-03-21